+++ crates/lib/examples/parametric_compilation.rs
@@ -1,4 +1,4 @@
-//! This example implements the [Parametric Compilation example from pyQuil][pyquil] in Rust.
+//! This example implements the [Parametric Compilation example from pyQuil](pyquil) in Rust.
 //! [pyquil]: https://pyquil-docs.rigetti.com/en/stable/basics.html?highlight=parametric#parametric-compilation
 
 use std::f64::consts::PI;
@@ -37,19 +37,20 @@ async fn main() {
             .duration
             .expect("Aspen-11 should always report duration");
 
-        let result = data
+        let ro_readout_data = data
             .readout_data
             .get_readout_values_for_field("ro")
-            .expect("Missing ro register")
-            .unwrap()
+            .expect("readout values should contain 'ro'")
+            .expect("'ro' should contain readout values");
+        let first_ro_data = ro_readout_data
             .first()
-            .unwrap()
-            .clone()
-            .unwrap()
+            .expect("'ro' should contain at least one readout value")
+            .clone();
+        let first_ro_values = first_ro_data
+            .expect("first readout value should ")
             .values
-            .unwrap();
-
-        if let Values::IntegerValues(IntegerReadoutValues { mut values }) = result {
+            .expect("'ro' should have readout values");
+        if let Values::IntegerValues(IntegerReadoutValues { mut values }) = first_ro_values {
             parametric_measurements.append(&mut values)
         }
     }