+++ crates/lib/src/qpu/client.rs
@@ -21,7 +21,7 @@ pub use qcs_api_client_common::configuration::LoadError;
 pub use qcs_api_client_grpc::channel::Error as GrpcError;
 pub use qcs_api_client_openapi::apis::Error as OpenApiError;
 
-#[derive(Clone)]
+#[derive(Clone, Default)]
 pub struct QcsClient {
     config: ClientConfiguration,
     /// When enabled, default to Gateway service for execution. Fallback to QPU's default endpoint otherwise.