+++ crates/lib/tests/parametric_compilation.rs
@@ -1,9 +1,10 @@
-//! This integration test implements the [Parametric Compilation example from pyQuil][example].
+//! This integration test implements the [Parametric Compilation example from pyQuil](example).
 //! [example]: https://pyquil-docs.rigetti.com/en/stable/basics.html?highlight=parametric#parametric-compilation
 
 use std::f64::consts::PI;
 
 use qcs::Executable;
+use qcs_api_client_openapi::common::ClientConfiguration;
 
 const BASIC_SUBSTITUTION: &str = r#"
 DECLARE ro BIT
@@ -18,7 +19,8 @@ MEASURE 0 ro[0]
 
 #[tokio::test]
 async fn basic_substitution() {
-    let mut exe = Executable::from_quil(BASIC_SUBSTITUTION);
+    let mut exe =
+        Executable::from_quil(BASIC_SUBSTITUTION).with_config(ClientConfiguration::default());
     let mut parametric_measurements = Vec::with_capacity(200);
 
     let step = 2.0 * PI / 200.0;