+++ crates/lib/CHANGELOG.md
@@ -1,3 +1,34 @@
+## 0.7.1-rc.22
+
+### Features
+
+- upgrade quil-rs dependency and fix issues (#172)
+- python bindings (#145)
+
+## 0.7.1-rc.21
+
+### Features
+
+- python bindings (#145)
+
+## 0.7.1-rc.20
+
+### Features
+
+- python bindings (#145)
+
+## 0.7.1-rc.19
+
+### Features
+
+- python bindings (#145)
+
+## 0.7.1-rc.18
+
+### Features
+
+- python bindings (#145)
+
 ## 0.7.1-rc.17
 
 ### Features