+++ crates/python/CHANGELOG.md
@@ -1,3 +1,39 @@
+## 0.2.1-rc.22
+
+### Features
+
+- upgrade quil-rs dependency and fix issues (#172)
+- add type hints
+- python bindings (#145)
+
+## 0.2.1-rc.21
+
+### Features
+
+- add type hints
+- python bindings (#145)
+
+## 0.2.1-rc.20
+
+### Features
+
+- add type hints
+- python bindings (#145)
+
+## 0.2.1-rc.19
+
+### Features
+
+- add type hints
+- python bindings (#145)
+
+## 0.2.1-rc.18
+
+### Features
+
+- add type hints
+- python bindings (#145)
+
 ## 0.2.1-rc.17
 
 ### Features